refactor(test_parse.py): Split render tests into their own file
