+++ adr_viewer/test_parse.py
@@ -1,4 +1,4 @@
-from adr_viewer import parse_adr_to_config, render_html
+from adr_viewer import parse_adr_to_config
 
 
 def test_should_extract_title_from_record():
@@ -42,48 +42,11 @@ def test_should_mark_pending_records():
 
     assert config['status'] == 'pending'
 
+
 def test_should_mark_pproposed_records():
     config = parse_adr_to_config('test/adr/0004-proposed-status.md')
 
     assert config['status'] == 'pending'
-
-def test_should_render_html_with_project_title():
-    html = render_html({
-        'project_title': 'my-project'
-    })
-
-    assert '<title>ADR Viewer - my-project</title>' in html
-
-
-def test_should_render_html_with_record_status():
-    html = render_html({
-        'records': [{
-            'status': 'accepted',
-        }]
-    })
-
-    assert '<div class="panel-heading adr-accepted">' in html
-
-
-def test_should_render_html_with_record_body():
-    html = render_html({
-        'records': [{
-            'body': '<h1>This is my ADR</h1>',
-        }]
-    })
-
-    assert '<div class="panel-body"><h1>This is my ADR</h1></div>' in html
-
-
-def test_should_render_html_with_collapsible_index():
-    html = render_html({
-        'records': [{
-            'title': 'Record 123',
-            'index': 123
-        }]
-    })
-
-    assert '<a data-toggle="collapse" href="#collapse123">Record 123</a>' in html
 
 
 def test_should_ignore_invalid_files():
